+++ src/vsix/UI/Events/EventManager.cs
@@ -1,57 +1,128 @@
-﻿using CommunityToolkit.Mvvm.Messaging;
+﻿using System;
 
 using Community.VisualStudio.Toolkit;
 
+using CommunityToolkit.Mvvm.Messaging;
+
 namespace StartPagePlus.UI.Events
 {
+    using Core;
+    using Core.Interfaces;
+
     using DI;
 
     using Messages;
 
+    using Options.Pages;
+
+    using Services;
+
     internal class EventManager
     {
-        /// <summary>
-        /// Initializes a new instance of the <see cref="EventManager"/> class.
-        /// </summary>
-        /// <remarks>
-        /// This constructor will produce an instance that will use the <see cref="WeakReferenceMessenger.Default"/> instance
-        /// to perform requested operations. It will also be available locally through the <see cref="Messenger"/> property.
-        /// </remarks>
-        static EventManager()
-        { Messenger = WeakReferenceMessenger.Default; }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="EventManager"/> class.
-        /// </summary>
-        /// <param name="messenger">The <see cref="IMessenger"/> instance to use to send messages.</param>
-        /// <exception cref="System.ArgumentNullException">Thrown if <paramref name="messenger"/> is <see langword="null"/>.</exception>
-        //static EventManager(IMessenger messenger)
-        //{
-        //    Messenger = messenger ?? throw new ArgumentNullException(nameof(messenger));
-        //}
+        private static readonly IDialogService _dialogService;
+        private static readonly IAsyncMethodService _methodService;
+        private static readonly IMessenger _messenger;
 
         //---
 
-        /// <summary>
-        /// Gets the <see cref="IMessenger"/> instance in use.
-        /// </summary>
-        private static IMessenger Messenger { get; }
+        static EventManager()
+        {
+            _dialogService = ServiceManager.DialogService;
+            _methodService = ServiceManager.AsyncMethodService;
+            _messenger = MessengerManager.Messenger;
+        }
 
         //---
 
         internal static void RegisterEvents(StartPagePlusContainer container)
         {
+            VS.Events.SolutionEvents.OnBeforeOpenSolution += OnBeforeOpenSolution;
             VS.Events.SolutionEvents.OnAfterCloseSolution += OnAfterCloseSolution;
         }
 
         internal static void DeregisterEvents(StartPagePlusContainer container)
         {
             VS.Events.SolutionEvents.OnAfterCloseSolution -= OnAfterCloseSolution;
+            VS.Events.SolutionEvents.OnBeforeOpenSolution -= OnBeforeOpenSolution;
         }
 
         //---
 
+        private static void OnBeforeOpenSolution(string obj)
+        {
+            if (GeneralOptions.Instance.HideOnSolutionOpen == true)
+            {
+                CloseWindow();
+            }
+
+            //---
+
+            static void CloseWindow()
+            {
+                _methodService.Run(
+                    async () =>
+                    {
+                        var result = false;
+
+                        try
+                        {
+                            var mainWindow = await VS.Windows.FindWindowAsync(PackageGuids.StartPagePlusWindow);
+
+                            await mainWindow.CloseFrameAsync(FrameCloseOption.NoSave);
+
+                            result = true;
+                        }
+                        catch (Exception ex)
+                        {
+                            result = false;
+                            _dialogService.ShowException(ex);
+                        }
+
+                        return result;
+                    }
+                );
+            }
+        }
+
         private static void OnAfterCloseSolution()
-            => Messenger.Send<RecentItemsRefresh>();
+        {
+            if (GeneralOptions.Instance.HideOnSolutionOpen == true)
+            {
+                ShowWindow();
+            }
+
+            RefreshRecentItems();
+
+            //---
+
+            static void ShowWindow()
+            {
+                _methodService.Run(
+                    async () =>
+                    {
+                        var result = false;
+
+                        try
+                        {
+                            var mainWindow = await VS.Windows.FindWindowAsync(PackageGuids.StartPagePlusWindow);
+
+                            await mainWindow.ShowAsync();
+
+                            result = true;
+                        }
+                        catch (Exception ex)
+                        {
+                            result = false;
+                            _dialogService.ShowException(ex);
+                        }
+
+                        return result;
+                    }
+                );
+            }
+
+            static void RefreshRecentItems()
+                => _messenger.Send<RecentItemsRefresh>();
+        }
     }
 }