+++ src/vsix/UI/Interfaces/IRunMethods.cs
@@ -9,12 +9,12 @@ namespace StartPagePlus.UI.Interfaces
         /// <remarks>Only use when the called code has an exception handler</remarks>
         /// <param name="asyncMethod">The asyncronous code to be run synchronously</param>
         /// <returns>A bool from the called code</returns>
-        bool RunMethod(Func<Task<bool>> asyncMethod);
+        bool Run(Func<Task<bool>> asyncMethod);
 
         /// <summary>A method to safely run asynchronous code syncronously</summary>
         /// <remarks>Only use when the called code has an exception handler</remarks>
         /// <param name="asyncMethod">The asyncronous code to be run synchronously</param>
         /// <returns>A bool? from the called code</returns>
-        bool? RunMethod(Func<Task<bool?>> asyncMethod);
+        bool? Run(Func<Task<bool?>> asyncMethod);
     }
 }