+++ src/vsix/UI/Interfaces/IViewModel.cs
@@ -1,5 +1,5 @@
 ﻿namespace StartPagePlus.UI.Interfaces
 {
-    public interface IViewModel : /*IMessageMethods,*/ IRunMethods
+    public interface IViewModel
     { }
 }